system metrics moved to shared_utils, improved slack outout
